ai_meme_generator_agent_browseruse/requirements.txt
openai>=1.0.0
langchain-openai>=0.2.0
browser-use>=0.1.0
python-dotenv>=1.0.0
